fix(qa): adopt split Matrix E2EE flows
