+++ extensions/qa-lab/src/live-transports/matrix/scenarios/scenario-runtime-e2ee-recovery.ts
@@ -1,8 +1,8 @@
 // Qa Matrix plugin module implements recovery scenario runtime E2EE behavior.
 import { randomUUID } from "node:crypto";
 import { truncateUtf16Safe } from "openclaw/plugin-sdk/text-utility-runtime";
-import { createMatrixQaClient } from "../../substrate/client.js";
-import { createMatrixQaE2eeScenarioClient } from "../../substrate/e2ee-client.js";
+import { createMatrixQaClient } from "../substrate/client.js";
+import { createMatrixQaE2eeScenarioClient } from "../substrate/e2ee-client.js";
 import {
   assertMatrixQaExpectedBootstrapFailure,
   assertMatrixQaFaultedRecoveryOwnerVerificationRequired,
