+++ qa/scenarios/channels/matrix-e2ee-cli-encryption-setup-multi-account.yaml
@@ -29,7 +29,7 @@ scenario:
             requireMention: true
 
 flow:
-  module: ./live-transports/matrix/scenarios/scenario-runtime-e2ee.js
+  module: ./live-transports/matrix/scenarios/scenario-runtime-e2ee-cli-gateway.js
   call: runMatrixQaE2eeCliEncryptionSetupMultiAccountScenario
   args:
     - expr: "scenarioContext"
