+++ qa/scenarios/channels/matrix-e2ee-cli-recovery-key-setup.yaml
@@ -29,7 +29,7 @@ scenario:
             requireMention: true
 
 flow:
-  module: ./live-transports/matrix/scenarios/scenario-runtime-e2ee.js
+  module: ./live-transports/matrix/scenarios/scenario-runtime-e2ee-cli-recovery.js
   call: runMatrixQaE2eeCliRecoveryKeySetupScenario
   args:
     - expr: "scenarioContext"
