+++ qa/scenarios/channels/matrix-e2ee-cli-setup-then-gateway-reply.yaml
@@ -29,7 +29,7 @@ scenario:
             requireMention: true
 
 flow:
-  module: ./live-transports/matrix/scenarios/scenario-runtime-e2ee.js
+  module: ./live-transports/matrix/scenarios/scenario-runtime-e2ee-cli-gateway.js
   call: runMatrixQaE2eeCliSetupThenGatewayReplyScenario
   args:
     - expr: "scenarioContext"
