+++ qa/scenarios/channels/matrix-e2ee-device-sas-verification.yaml
@@ -29,7 +29,7 @@ scenario:
             requireMention: true
 
 flow:
-  module: ./live-transports/matrix/scenarios/scenario-runtime-e2ee.js
+  module: ./live-transports/matrix/scenarios/scenario-runtime-e2ee-verification.js
   call: runMatrixQaE2eeDeviceSasVerificationScenario
   args:
     - expr: "scenarioContext"
