+++ qa/scenarios/channels/matrix-e2ee-recovery-owner-verification-required.yaml
@@ -29,7 +29,7 @@ scenario:
             requireMention: true
 
 flow:
-  module: ./live-transports/matrix/scenarios/scenario-runtime-e2ee.js
+  module: ./live-transports/matrix/scenarios/scenario-runtime-e2ee-recovery.js
   call: runMatrixQaE2eeRecoveryOwnerVerificationRequiredScenario
   args:
     - expr: "scenarioContext"
