+++ qa/scenarios/channels/matrix-e2ee-stale-device-hygiene.yaml
@@ -29,7 +29,7 @@ scenario:
             requireMention: true
 
 flow:
-  module: ./live-transports/matrix/scenarios/scenario-runtime-e2ee.js
+  module: ./live-transports/matrix/scenarios/scenario-runtime-e2ee-verification.js
   call: runMatrixQaE2eeStaleDeviceHygieneScenario
   args:
     - expr: "scenarioContext"
